Attempt to fix the logic used to close Outline if it is the last window
The previous logic that got triggered when the main buffer to which
Outline is attached gets closed was closing the Outline but left the
main buffer as it is. If I open neo-tree on startup (or even later)
Outline and the main buffer closes correctly when the main buffer is
closed. I tried removing neo-tree setup but that didn't help close the
main buffer - still only the Outline was closing when the main buffer
was closed.

Now I changed the logic to be the same one as vim-ccls used to show call
hierarchy. When Outline or call hierarchy is the last window, it will
receive focus. At this point if we see that there is no main buffer, we
can quit vim.